+++ fuel-core/src/query/output_proof/test.rs
